% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateParamChoices}
\alias{generateParamChoices}
\title{Generate a dataframe with parameter options for model creation.}
\usage{
generateParamChoices(predictVariables, modelReturn, pathToSave = "",
  save = FALSE)
}
\arguments{
\item{predictVariables}{string vector of prediction variables}

\item{modelReturn}{censReg object returned from censReg}

\item{pathToSave}{string}

\item{save}{logical}
}
\description{
Generate a dataframe and optional csv file to be used to narrow down allowed parameters
in model creation. Creates a dataframe in which one column
is named 'Scalar', one is named 'variableNames', and the rest are the available parameters.
Each row in the dataframe contains one of the available parameters. Using ones and zeros,
formulas can easily be created including log transforms and interaction terms with the \link{createFormulaFromDF} function.
}
\examples{
DTComplete <- StLouisDT
UV <- StLouisUV
response <- "Ammonia.N"
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
predictVariables <- names(DT)[-which(names(DT) \%in\% c(response,"datetime","decYear"))]
kitchenSink <- createFullFormula(DT,response)
returnPrelim <- prelimModelDev(DT,response,kitchenSink)
modelReturn <- returnPrelim$DT.mod
choices <- generateParamChoices(predictVariables,modelReturn)
}
\keyword{list}
\keyword{model}
\keyword{parameter}

