% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mergeDatasets.R
\name{mergeDatasets}
\alias{mergeDatasets}
\title{Creates a dataframe of merged datasets.}
\usage{
mergeDatasets(localQW, localUV, QWcodes, max.diff = "2 hours")
}
\arguments{
\item{localQW}{dataframe water quality dataset. Should have qw objects,
along with a column called "datetime" in POSIXct.}

\item{localUV}{dataframe unit value dataset. Should have one column called "datetime" in POSIXct.}

\item{QWcodes}{dataframe}

\item{max.diff}{string default is "2 hours". See \code{?mergeNearest} for more options.}
}
\value{
retList list including the DT list and updated QWcodes
}
\description{
Creates a list of merged datasets, along with updating the QWcodes dataframe. The function looks for
qw objects in the QW dataframe, cross-references them with the QWcodes table, and merges those
columns with the nearest UV (unit value) data. The function will merge the UV data that is nearest
to the QW data, see \code{?mergeNearest} for more options on specifying the acceptable time gap between
the two datasets.
}
\examples{
UV <- StLouisUV
QW <- StLouisQW
QWcodes <- StLouisQWcodes
mergeReturn <- mergeDatasets(QW, UV, QWcodes)
DTComplete <- mergeReturn$DTComplete
QWcodes <- mergeReturn$QWcodes
}
\keyword{datasets}
\keyword{merge}

