% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotQQTransforms.R
\name{plotQQTransforms}
\alias{plotQQTransforms}
\title{Plots Q-Q of predict variables.}
\usage{
plotQQTransforms(localDT, responseVariable, printLinear = TRUE,
  printLog = TRUE, printSquare = FALSE, printSQRT = FALSE,
  printEXP = FALSE)
}
\arguments{
\item{localDT}{dataframe in wide format}

\item{responseVariable}{character}

\item{printLinear}{logical option to print a page of linear plots (default to TRUE)}

\item{printLog}{logical option to print a page of log (base 10) plots (defaults to TRUE)}

\item{printSquare}{logical option to print a page of squared plots (defaults to FALSE)}

\item{printSQRT}{logical option to print a page of square root plots (defaults to FALSE)}

\item{printEXP}{logical option to print a page of exponential plots (defaults to FALSE)}
}
\description{
Plots Q-Q of predict variables.See \link{qqmath} for more information on q-q (quantile-quantile) plots.
}
\examples{
DTComplete <- StLouisDT
colnames(DTComplete) <- gsub("_Inst","",colnames(DTComplete))
UV <- StLouisUV
colnames(UV) <- gsub("_Inst","",colnames(UV))
response <- "Ammonia.N"
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
plotQQTransforms(DT,response)
}
\keyword{qq}

