% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resultResidPlots}
\alias{resultResidPlots}
\title{Plots model summary plots using residuals}
\usage{
resultResidPlots(localDT, finalModel, siteINFO)
}
\arguments{
\item{localDT}{dataframe in wide format}

\item{finalModel}{censReg model results}

\item{siteINFO}{dataframe including station name (station.nm) and siteID (site.no) (easiestly retrieved from dataRetrieval package)}
}
\value{
plot
}
\description{
Plots a set of plots to summarize the model results and behavior. The first plot (A) is always observed
versus predicted, the second (B) is residuals vs. predicted, the third (C) is residuals vs. time, and
the fourth (D) is residual quantiles vs. theoretical quantiles. There are then individual plots for each
prediction variable of residuals vs. prediction value. Residuals for censored values are calculated
using the detection limit.
}
\examples{
DTComplete <- StLouisDT
UV <- StLouisUV
response <- "Ammonia.N"
siteINFO <- StLouisInfo
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
kitchenSink <- createFullFormula(DT,response)
returnPrelim <- prelimModelDev(DT,response,kitchenSink)
modelReturn <- returnPrelim$DT.mod
resultResidPlots(DT,modelReturn,siteINFO)
}
\keyword{scatterplot}

