% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rSqr.R
\name{rSqr}
\alias{rSqr}
\title{Calculates R^2 and adjusted R^2 values}
\usage{
rSqr(observed, resid, predicted)
}
\arguments{
\item{observed}{number vector of observed data}

\item{resid}{number vector of residual data}

\item{predicted}{number vector of predicted data}
}
\value{
df list with rSquare and adjusted rSquare
}
\description{
Calculates R^2 and adjusted R^2 values
}
\examples{
observed <- runif(10, 5.0, 7.5)
resid <- runif(10, 0.1, 0.2)
predicted<-1
rList <- rSqr(observed, resid, predicted)
}
\keyword{rSquared}

