% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saveModelResults.R
\name{saveModelResults}
\alias{saveModelResults}
\title{Saves critical objects for model analysis}
\usage{
saveModelResults(pathToSave, DT, UV, modelReturn, siteINFO)
}
\arguments{
\item{pathToSave}{full path for saving results}

\item{DT}{dataframe of potential input variables to model}

\item{UV}{dataframe of unit values}

\item{modelReturn}{results from censReg}

\item{siteINFO}{dataframe that includes station_nm and site.no}
}
\description{
Saves critical objects for model analysis
}
\examples{
\dontrun{}
}
\keyword{model}
\keyword{results}
\keyword{save}

