% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{Return string with initial capital letters}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{string}
}
\value{
retVal corrected string
}
\description{
Return string with initial capital letters
}
\examples{
exString <- "BAD RIVER"
exString <- simpleCap(exString)
}
\keyword{formatting}
\keyword{string}

