% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/view.R
\name{View}
\alias{View}
\title{View QW objects}
\usage{
View(x)
}
\arguments{
\item{x}{dataframe to view.}
}
\description{
View QW objects, specifically a fix for RStudio.
}
\keyword{custom}
\keyword{view}

