% Generated by roxygen2 (4.0.2): do not edit by hand
\name{whatUVNew}
\alias{whatUVNew}
\title{Imports a table of available parameters, period of record, and count.}
\usage{
whatUVNew(siteNumber, service = "uv")
}
\arguments{
\item{siteNumber}{string USGS site number.  This is usually an 8 digit number}

\item{service}{string defaults to "uv" (unit values), other options are "qw" (water quality) and "dv" (daily values)}
}
\value{
retval dataframe with all information found in the expanded site file
}
\description{
Imports a table of available parameters, period of record, and count. Information is retrieved from \url{http://waterservices.usgs.gov/}
site information service.
}
\examples{
site <- "04027000"
uv <-whatUVNew(site)
}
\keyword{USGS}
\keyword{data}
\keyword{import}
\keyword{service}
\keyword{web}

