% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createFullDT.R
\name{createFullDT}
\alias{createFullDT}
\title{Create log and interaction columns}
\usage{
createFullDT(formulaToUse, localDT)
}
\arguments{
\item{formulaToUse}{formula}

\item{localDT}{dataframe of potential input variables to model}
}
\value{
fullDTList list of DT (data frame with new log and interaction columns),
modelFormula (text of formula to use with new column names),
and colNames (all column names)
}
\description{
Create log and interaction columns. Using a formula and raw data, log and interaction columns will be computed.
If negative values are found in the raw data column that need to be log transformed, the values are changed to 1.
}
\examples{
DTComplete <- StLouisDT
colnames(DTComplete) <- gsub("_Inst","",colnames(DTComplete))
UV <- StLouisUV
colnames(UV) <- gsub("_Inst","",colnames(UV))
response <- "Ammonia.N"
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
basicFormula <- "Flow + log(Flow) + Wtemp:Flow"
returnData <- createFullDT(basicFormula,DT)
newDT <- returnData$DT
modelFormula <- returnData$modelFormula
newColumns <- returnData$colNames
}
\keyword{transforms}

