% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/externalStudentRes.R
\name{externalStudentRes}
\alias{externalStudentRes}
\title{Compute internally Studentized residuals}
\usage{
externalStudentRes(localDT, modelCoefList, transformResponse = "lognormal")
}
\arguments{
\item{localDT}{dataframe that includes all response and predictor variables}

\item{modelCoefList}{list of model coefficients}

\item{transformResponse}{string can be "normal" or "lognormal", perhaps try to generalize this more in future}
}
\value{
StRes.all
}
\description{
Compute internally Studentized residuals. The function can calculate multiple models, therefore
the model coefficients most be input as a named list (where the name is the response variable for
that model result). Residuals of censored values are calculated simply by the prediction minus detection limit.
Predictions are calculated with the \link{censReg} function.
}
\examples{
DTComplete <- StLouisDT
UV <- StLouisUV
response <- "Ammonia.N"
siteINFO <- StLouisInfo
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
kitchenSink <- createFullFormula(DT,response)
returnPrelim <- prelimModelDev(DT,response,kitchenSink)
modelReturn <- returnPrelim$DT.mod
modelCoef <- modelReturn$PARAML
names(modelCoef) <- c(dimnames(modelReturn$XLCAL)[[2]],"logSigma")
modelCoefList <- list()
modelCoefList[[rownames(attributes(modelReturn$terms)$factors)[1]]] <- modelCoef
extStudent <- externalStudentRes(DT,modelCoefList)
}
\keyword{residuals}
\keyword{studentized}

