% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findOutliers.R
\name{findOutliers}
\alias{findOutliers}
\title{Calculates potential outliers based on external studentized residuals}
\usage{
findOutliers(modelReturn, localDT, transformResponse = "lognormal")
}
\arguments{
\item{modelReturn}{list returned from censReg}

\item{localDT}{DTframe that includes all response and predictor variables}

\item{transformResponse}{string can be "normal" or "lognormal", perhaps try to generalize this more in future}
}
\value{
outlier vector of index numbers
}
\description{
Find index of outliers using external studentized residuals. Outliers are values that have
external studentized residuals greater than 3 or less than negative 3.
}
\examples{
DTComplete <- StLouisDT
colnames(DTComplete) <- gsub("_Inst","",colnames(DTComplete))
UV <- StLouisUV
colnames(UV) <- gsub("_Inst","",colnames(UV))
QWcodes <- StLouisQWcodes
siteINFO <- StLouisInfo
response <- QWcodes$colName[QWcodes$colName != ""][1]
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
kitchenSink <- createFullFormula(DT,response)
returnPrelim <- prelimModelDev(DT,response,kitchenSink)
modelReturn <- returnPrelim$DT.mod
outlierIndex <- findOutliers(modelReturn,DT)
}
\keyword{residuals}
\keyword{studentized}

