% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPredictVariables}
\alias{getPredictVariables}
\title{Returns a character vector of prediction variables.}
\usage{
getPredictVariables(DTnames)
}
\arguments{
\item{DTnames}{column names of DT dataframe}
}
\value{
predictVariables string predict variables based on column headers
}
\description{
Returns a character vector of prediction variables. Accepts column names as an input, this function
filters out '_cd' columns as well as "agency", "site", "tz". These are typical column names
retrieved from NWIS web services.
}
\examples{
UV <- StLouisUV
predictVars <- getPredictVariables(names(UV))
}
\keyword{predict}

