% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSteps.R
\name{plotSteps}
\alias{plotSteps}
\title{Plots showing obs. vs pred. of each step in stepwise regression}
\usage{
plotSteps(steps, localDT, transformResponse = "lognormal")
}
\arguments{
\item{steps}{dataframe}

\item{localDT}{dataframe of potential input variables to model}

\item{transformResponse}{ring can be "normal" or "lognormal", perhaps try to generalize this more in future}
}
\value{
retVal list of modelStuff, steps, DT.mod
}
\description{
Plots showing observed versus predicted for each step calculated in the stepwise
regression. A 1:1 blue line is plotted for reference. The results of a basic linear regression
between all the points is shown in red. Censored values are signified with a segment from the detection level.
Outliers are shown as solid red circles, as calculated by \code{findOutliers}. In this function,
outliers are defined as points that have greater than 3 external studentized residuals.
}
\examples{
DTComplete <- StLouisDT
UV <- StLouisUV
response <- "Ammonia.N"
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
kitchenSink <- createFullFormula(DT,response)
returnPrelim <- prelimModelDev(DT,response,kitchenSink)
steps <- returnPrelim$steps
plotSteps(steps,DT)
}
\keyword{plot}

