% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predictVariableScatterPlots.R
\name{predictVariableScatterPlots}
\alias{predictVariableScatterPlots}
\title{Plots scatterplot of predict variables.}
\usage{
predictVariableScatterPlots(localDT, responseVariable,
  transformResponse = "lognormal")
}
\arguments{
\item{localDT}{dataframe in wide format}

\item{responseVariable}{string column name of response variable in localDT dataframe}

\item{transformResponse}{string can be "normal" or "lognormal", perhaps try to generalize this more in future}
}
\description{
Plots scatterplot of predict variables. One page of linear, and one page of log transformed data.
}
\examples{
DTComplete <- StLouisDT
colnames(DTComplete) <- gsub("_Inst","",colnames(DTComplete))
UV <- StLouisUV
colnames(UV) <- gsub("_Inst","",colnames(UV))
response <- "Ammonia.N"
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
predictVariableScatterPlots(DT,response)
}
\keyword{scatterplot}

