% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resultPlots.R
\name{resultPlots}
\alias{resultPlots}
\title{Plots model summary plots}
\usage{
resultPlots(localDT, finalModel, siteINFO)
}
\arguments{
\item{localDT}{dataframe in wide format}

\item{finalModel}{censReg model results}

\item{siteINFO}{dataframe including station name (station_nm) and siteID (site.no) (easiestly retrieved from dataRetrieval package)}
}
\value{
plot
}
\description{
Plots a set of plots to summarize the model results and behavior. The first plot (A) is always observed
versus predicted, the second (B) is residuals vs. predicted, the third (C) is residuals vs. time, and
the fourth (D) is residual quantiles vs. theoretical quantiles. There are then individual plots for each
prediction variable of observed vs. prediction value. Residuals for censored values are calculated
using the detection limit.
}
\examples{
DTComplete <- StLouisDT
colnames(DTComplete) <- gsub("_Inst","",colnames(DTComplete))
UV <- StLouisUV
colnames(UV) <- gsub("_Inst","",colnames(UV))
response <- "Ammonia.N"
siteINFO <- StLouisInfo
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
kitchenSink <- createFullFormula(DT,response)
returnPrelim <- prelimModelDev(DT,response,kitchenSink)
modelReturn <- returnPrelim$DT.mod
resultPlots(DT,modelReturn,siteINFO)
}
\keyword{scatterplot}

