% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runFit}
\alias{runFit}
\title{Run the fitting routine}
\usage{
runFit(localDT, finalModel, transformResponse = "lognormal")
}
\arguments{
\item{localDT}{dataframe of potential input variables to model}

\item{finalModel}{results from censReg}

\item{transformResponse}{string can be "normal" or "lognormal", perhaps try to generalize this more in future}
}
\value{
evaluat list
}
\description{
Run the fitting routine
}
\examples{
DTComplete <- StLouisDT
UV <- StLouisUV
response <- "Ammonia.N"
siteINFO <- StLouisInfo
DT <- DTComplete[c(response,getPredictVariables(names(UV)), "decYear","sinDY","cosDY","datetime")]
DT <- na.omit(DT)
kitchenSink <- createFullFormula(DT,response)
returnPrelim <- prelimModelDev(DT,response,kitchenSink)
modelReturn <- returnPrelim$DT.mod
fitReturn <- runFit(DT,modelReturn)
}
\keyword{AMLE}
\keyword{fit}

